+++ src/index.js
@@ -43,7 +43,7 @@ app.use("/auth", require("./routes/auth.routes"))
 app.use("/user", require("./routes/user.routes"))
 app.use("/file", require("./routes/file.routes"))
 app.use("/search", require("./routes/search.routes"))
-app.use("/post", require("./routes/post.routes"))
+app.use("/pool", require("./routes/pool.routes"))
 app.use("/comment", require("./routes/comment.routes"))
 app.use("/view", require("./routes/view.routes"))
 app.use("/vote", require("./routes/vote.routes"))
